tests/integrations/sklearn/requirements.txt
scikit-learn==0.24.1
neptune-client==0.9.4
neptune-sklearn==0.9.1